@@ -1,7 +1,7 @@
 {
   "name": "@therockstorm/utils",
-  "version": "2.1.0",
-  "description": "Common utility functions",
+  "version": "3.0.3",
+  "description": "Tiny utility functions",
   "homepage": "https://github.com/therockstorm/utils#readme",
   "bugs": {
     "url": "https://github.com/therockstorm/utils/issues"
@@ -14,29 +14,29 @@
   "files": [
     "lib"
   ],
-  "main": "lib/index.js",
+  "main": "lib/src/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/therockstorm/utils.git"
   },
   "scripts": {
-    "build": "rm -rf lib && tsc --outDir lib && cp src/index.d.ts lib",
+    "build": "rm -rf lib && tsc --outDir lib && cp src/index.d.ts lib/src",
     "format": "prettier --write '{,!(lib)/**/}*.+(js|json|md|ts)'",
     "prepare": "npm run build",
     "pretest": "npm run format",
+    "start": "tsc && node src/index.js",
     "test": "mocha -r ts-node/register test/**/*.ts",
     "watch": "npm test -- --watch"
   },
   "devDependencies": {
     "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.14",
-    "husky": "^1.2.0",
-    "lint-staged": "^8.1.0",
+    "@types/node": "^11.9.3",
+    "husky": "^1.3.1",
+    "lint-staged": "^8.1.3",
     "mocha": "^5.2.0",
-    "prettier": "^1.15.3",
-    "sinon": "^7.2.2",
-    "ts-node": "^7.0.1",
-    "typescript": "^3.2.2"
+    "prettier": "^1.16.4",
+    "ts-node": "^8.0.2",
+    "typescript": "^3.3.3"
   },
   "husky": {
     "hooks": {
